requirements-dev.txt
-r requirements.txt

django-rosetta==0.9.8
flake8==6.0.0
python-dotenv==0.21.1
black==23.1.0